feat: refactor history card components
